+++ week4/package.json
@@ -11,7 +11,8 @@
     "axios": "^0.27.2",
     "json-server": "^0.17.0",
     "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "react-dom": "^18.0.0",
+    "styled-components": "^5.3.5"
   },
   "devDependencies": {
     "@types/react": "^18.0.0",
